kafka/consumer/requirements.txt
kafka-python==2.0.2
pandas==2.2.3
pyarrow==17.0.0
google-cloud-storage==2.18.2
